requirements.txt
python-telegram-bot>=20.7,<21.0
aiosqlite>=0.19
python-dotenv>=1.0

